Created dynamic lib containing c functions, Task 2

+++ 0x18-dynamic_libraries/c-math-functions.c
@@ -6,7 +6,7 @@ int add(int a, int b)
 	return (a + b);
 }
 // Subtraction
-int sub(double a, double b)
+int sub(int a, int b)
 {
 	return (a - b);
 }
